+++ src/main/java/com/sparta/startup_be/dto/GoogleUserInfoDto.java
@@ -1,0 +1,11 @@
+//package com.sparta.startup_be.dto;
+//
+//import lombok.Getter;
+//import lombok.NoArgsConstructor;
+//
+//@Getter
+//@NoArgsConstructor
+//public class GoogleUserInfoDto {
+//    private
+//
+//}
