+++ src/main/java/com/sparta/startup_be/security/WebSecurityConfig.java
@@ -134,10 +134,10 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 
         // 회원 관리 API 허용
         skipPathList.add("GET,/user/**");
-        skipPathList.add("GET,**");
-        skipPathList.add("POST,**");
-        skipPathList.add("PUT,**");
-        skipPathList.add("DELETE,**");
+//        skipPathList.add("GET,**");
+//        skipPathList.add("POST,**");
+//        skipPathList.add("PUT,**");
+//        skipPathList.add("DELETE,**");
 
         skipPathList.add("POST,/user/signup");
         skipPathList.add("POST,/user/login");
