+++ src/main/java/com/sparta/startup_be/security/filter/FormLoginFilter.java
@@ -26,6 +26,7 @@ public class FormLoginFilter extends UsernamePasswordAuthenticationFilter {
         UsernamePasswordAuthenticationToken authRequest;
         try {
             JsonNode requestBody = objectMapper.readTree(request.getInputStream());
+            System.out.println(requestBody.toString());
             String username = requestBody.get("userEmail").asText();
             String password = requestBody.get("password").asText();
             authRequest = new UsernamePasswordAuthenticationToken(username, password);
