+++ src/main/resources/application.properties
@@ -2,5 +2,5 @@ spring.h2.console.enabled=true
 spring.datasource.url=jdbc:h2:mem:testdb
 spring.datasource.username=sa
 spring.datasource.password=
-#spring.h2.console.settings.web-allow-others=true
-#spring.datasource.driver-class-name=org.h2.Driver
+spring.h2.console.settings.web-allow-others=true
+spring.datasource.driver-class-name=org.h2.Driver
